remove the header file and access the image dynamically
